Move 'is' to private folder
